+++ src/api/mock_gatewayer_test.go
@@ -394,6 +394,29 @@ func (_m *MockGatewayer) GetConnection(addr string) (*daemon.Connection, error) 
 	return r0, r1
 }
 
+// GetConnections provides a mock function with given fields:
+func (_m *MockGatewayer) GetConnections() ([]daemon.Connection, error) {
+	ret := _m.Called()
+
+	var r0 []daemon.Connection
+	if rf, ok := ret.Get(0).(func() []daemon.Connection); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).([]daemon.Connection)
+		}
+	}
+
+	var r1 error
+	if rf, ok := ret.Get(1).(func() error); ok {
+		r1 = rf()
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
 // GetDefaultConnections provides a mock function with given fields:
 func (_m *MockGatewayer) GetDefaultConnections() []string {
 	ret := _m.Called()
@@ -436,6 +459,29 @@ func (_m *MockGatewayer) GetHealth() (*daemon.Health, error) {
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(*daemon.Health)
+		}
+	}
+
+	var r1 error
+	if rf, ok := ret.Get(1).(func() error); ok {
+		r1 = rf()
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
+// GetIncomingConnections provides a mock function with given fields:
+func (_m *MockGatewayer) GetIncomingConnections() ([]daemon.Connection, error) {
+	ret := _m.Called()
+
+	var r0 []daemon.Connection
+	if rf, ok := ret.Get(0).(func() []daemon.Connection); ok {
+		r0 = rf()
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).([]daemon.Connection)
 		}
 	}