+++ src/api/network_test.go
@@ -126,13 +126,18 @@ func TestConnection(t *testing.T) {
 
 func TestConnections(t *testing.T) {
 	tt := []struct {
-		name                                 string
-		method                               string
-		status                               int
-		err                                  string
-		gatewayGetSolicitedConnectionsResult []daemon.Connection
-		gatewayGetSolicitedConnectionsError  error
-		result                               Connections
+		name                                string
+		method                              string
+		typ                                 string
+		status                              int
+		err                                 string
+		gatewayGetConnectionsResult         []daemon.Connection
+		gatewayGetConnectionsError          error
+		gatewayGetOutgoingConnectionsResult []daemon.Connection
+		gatewayGetOutgoingConnectionsError  error
+		gatewayGetIncomingConnectionsResult []daemon.Connection
+		gatewayGetIncomingConnectionsError  error
+		result                              Connections
 	}{
 		{
 			name:   "405",
@@ -140,12 +145,79 @@ func TestConnections(t *testing.T) {
 			status: http.StatusMethodNotAllowed,
 			err:    "405 Method Not Allowed",
 		},
+
 		{
-			name:   "200",
+			name:   "400 bad type",
+			method: http.MethodGet,
+			typ:    "foo",
+			status: http.StatusBadRequest,
+			err:    "400 Bad Request - invalid type",
+		},
+
+		{
+			name:   "200 all connections",
 			method: http.MethodGet,
 			status: http.StatusOK,
 			err:    "",
-			gatewayGetSolicitedConnectionsResult: []daemon.Connection{
+			gatewayGetConnectionsResult: []daemon.Connection{
+				{
+					ID:           1,
+					Addr:         "127.0.0.1",
+					LastSent:     99999,
+					LastReceived: 1111111,
+					Outgoing:     true,
+					Introduced:   true,
+					Mirror:       9876,
+					ListenPort:   9877,
+					Height:       1234,
+				},
+				{
+					ID:           3,
+					Addr:         "127.1.1.1",
+					LastSent:     88888,
+					LastReceived: 9999999,
+					Outgoing:     false,
+					Introduced:   true,
+					Mirror:       9877,
+					ListenPort:   9877,
+					Height:       1235,
+				},
+			},
+			result: Connections{
+				Connections: []readable.Connection{
+					{
+						ID:           1,
+						Addr:         "127.0.0.1",
+						LastSent:     99999,
+						LastReceived: 1111111,
+						Outgoing:     true,
+						Introduced:   true,
+						Mirror:       9876,
+						ListenPort:   9877,
+						Height:       1234,
+					},
+					{
+						ID:           3,
+						Addr:         "127.1.1.1",
+						LastSent:     88888,
+						LastReceived: 9999999,
+						Outgoing:     false,
+						Introduced:   true,
+						Mirror:       9877,
+						ListenPort:   9877,
+						Height:       1235,
+					},
+				},
+			},
+		},
+
+		{
+			name:   "200 outgoing connections",
+			method: http.MethodGet,
+			typ:    "outgoing",
+			status: http.StatusOK,
+			err:    "",
+			gatewayGetOutgoingConnectionsResult: []daemon.Connection{
 				{
 					ID:           1,
 					Addr:         "127.0.0.1",
@@ -176,18 +248,85 @@ func TestConnections(t *testing.T) {
 		},
 
 		{
-			name:                                "500 - GetOutgoingConnections failed",
-			method:                              http.MethodGet,
-			status:                              http.StatusInternalServerError,
-			err:                                 "500 Internal Server Error - GetOutgoingConnections failed",
-			gatewayGetSolicitedConnectionsError: errors.New("GetOutgoingConnections failed"),
+			name:   "200 incoming connections",
+			method: http.MethodGet,
+			typ:    "incoming",
+			status: http.StatusOK,
+			err:    "",
+			gatewayGetIncomingConnectionsResult: []daemon.Connection{
+				{
+					ID:           1,
+					Addr:         "127.0.0.1",
+					LastSent:     99999,
+					LastReceived: 1111111,
+					Outgoing:     false,
+					Introduced:   true,
+					Mirror:       9876,
+					ListenPort:   9877,
+					Height:       1234,
+				},
+			},
+			result: Connections{
+				Connections: []readable.Connection{
+					{
+						ID:           1,
+						Addr:         "127.0.0.1",
+						LastSent:     99999,
+						LastReceived: 1111111,
+						Outgoing:     false,
+						Introduced:   true,
+						Mirror:       9876,
+						ListenPort:   9877,
+						Height:       1234,
+					},
+				},
+			},
+		},
+
+		{
+			name:                       "500 - GetConnections failed",
+			method:                     http.MethodGet,
+			status:                     http.StatusInternalServerError,
+			err:                        "500 Internal Server Error - GetConnections failed",
+			gatewayGetConnectionsError: errors.New("GetConnections failed"),
+		},
+
+		{
+			name:                               "500 - GetOutgoingConnections failed",
+			method:                             http.MethodGet,
+			typ:                                "outgoing",
+			status:                             http.StatusInternalServerError,
+			err:                                "500 Internal Server Error - GetOutgoingConnections failed",
+			gatewayGetOutgoingConnectionsError: errors.New("GetOutgoingConnections failed"),
+		},
+
+		{
+			name:                               "500 - GetIncomingConnections failed",
+			method:                             http.MethodGet,
+			typ:                                "incoming",
+			status:                             http.StatusInternalServerError,
+			err:                                "500 Internal Server Error - GetIncomingConnections failed",
+			gatewayGetIncomingConnectionsError: errors.New("GetIncomingConnections failed"),
 		},
 	}
+
 	for _, tc := range tt {
 		t.Run(tc.name, func(t *testing.T) {
 			endpoint := "/api/v1/network/connections"
 			gateway := &MockGatewayer{}
-			gateway.On("GetOutgoingConnections").Return(tc.gatewayGetSolicitedConnectionsResult, tc.gatewayGetSolicitedConnectionsError)
+			gateway.On("GetConnections").Return(tc.gatewayGetConnectionsResult, tc.gatewayGetConnectionsError)
+			gateway.On("GetOutgoingConnections").Return(tc.gatewayGetOutgoingConnectionsResult, tc.gatewayGetOutgoingConnectionsError)
+			gateway.On("GetIncomingConnections").Return(tc.gatewayGetIncomingConnectionsResult, tc.gatewayGetIncomingConnectionsError)
+
+			v := url.Values{}
+			if tc.typ != "" {
+				v.Add("type", tc.typ)
+			}
+
+			ve := v.Encode()
+			if ve != "" {
+				endpoint += "?" + ve
+			}
 
 			req, err := http.NewRequest(tc.method, endpoint, nil)
 			require.NoError(t, err)